requirements.txt
streamlit
keras
tensorflow
pillow
pillow-heif